Links for HATEOS inserted
Links are included in the Task class. Still missed for the Collection.

The header Injections should be repleaced with Injections.

To note that the Links implementation is provided by Jersey with proprietary
APIs and this has been adopted instead of a more generic custom implementations
of the link. To move to a different REST Framework supporting JAX-RS this part
should be stripped and implemented according to the framework.