Get task details
Implemented the GET for the task. Addtionally, replaced the autoincrement ids for the task
with a string id initialised with a random UUID.

This allow to use user assigned ID for the task with the only problem of avoid conflicts
during the creation by manage the exception for the self generated or return the error
code to the user in case of user generated id.