Added Jersey Test Framework
Tests are not emabled by default and require the maven profile
tests. Tests run on an external container, which is tomcat, because
the inmemory container, and the other options, have problems to support
the JPA.

To run the test the package has to be built and deployed in a container
running in localhost at port 8080.