Completed test integration
Test integration works properly but the model could be improved
by creating separate methods for different tests. Currently,
tests are performed in group.

Separation could simplify the identification of the problem.

Unit tests are not performed because the code is tangled with JPA
and other service oriented features. Unit tests will be written later
if needed for the operation but very complicate for this part of code.